+++ lib/src/query_result.dart
@@ -2,46 +2,47 @@ import 'dart:collection';
 
 import 'query.dart';
 
-class PostgreSQLQueryResult extends UnmodifiableListView<PostgreSQLRow> {
-  final PostgreSQLQueryMetaData metaData;
+export 'query.dart' show PostgreSQLResultColumn;
 
-  PostgreSQLQueryResult(this.metaData, List<PostgreSQLRow> rows) : super(rows);
+class PostgreSQLResult extends UnmodifiableListView<PostgreSQLResultRow> {
+  final PostgreSQLResultMetaData metaData;
+
+  PostgreSQLResult(this.metaData, List<PostgreSQLResultRow> rows) : super(rows);
 }
 
-class PostgreSQLQueryMetaData {
-  final List<FieldDescription> fieldDescriptions;
+class PostgreSQLResultMetaData {
+  final List<PostgreSQLResultColumn> columns;
   List<String> _tableNames;
 
-  PostgreSQLQueryMetaData({this.fieldDescriptions});
+  PostgreSQLResultMetaData({this.columns});
 
   List<String> get tableNames {
-    _tableNames ??= fieldDescriptions
-        .map((column) => column.resolvedTableName)
-        .toSet()
-        .toList();
+    _tableNames ??=
+        columns.map((column) => column.resolvedTableName).toSet().toList();
     return _tableNames;
   }
 }
 
-class PostgreSQLRow extends UnmodifiableListView {
-  final PostgreSQLQueryMetaData metaData;
+class PostgreSQLResultRow extends UnmodifiableListView {
+  final PostgreSQLResultMetaData metaData;
   Map<String, Map<String, dynamic>> _tableMap;
 
-  PostgreSQLRow(this.metaData, List columns) : super(columns);
+  PostgreSQLResultRow(this.metaData, List columns) : super(columns);
 
+  /// Returns a two-level map that on the first level contains the resolved
+  /// table name, and on the second level the column name (or its alias).
   Map<String, Map<String, dynamic>> toTableMap() {
     if (_tableMap == null) {
-      final columns = metaData.fieldDescriptions;
       _tableMap = Map<String, Map<String, dynamic>>.fromIterable(
           metaData.tableNames,
           key: (name) => name as String,
           value: (_) => <String, dynamic>{});
 
-      final iterator = columns.iterator;
+      final iterator = metaData.columns.iterator;
       forEach((column) {
         iterator.moveNext();
-        _tableMap[iterator.current.resolvedTableName]
-            [iterator.current.fieldName] = column;
+        final col = iterator.current;
+        _tableMap[col.resolvedTableName][col.name] = column;
       });
     }
     return _tableMap;