+++ test/map_return_test.dart
@@ -147,7 +147,7 @@ class InterceptingConnection extends PostgreSQLConnection {
   List<String> queries = [];
 
   @override
-  Future<PostgreSQLQueryResult> query(String fmtString,
+  Future<PostgreSQLResult> query(String fmtString,
       {Map<String, dynamic> substitutionValues,
       bool allowReuse = true,
       int timeoutInSeconds}) {